added @Query("type") to GeocodingApiService.kt

+++ app/src/main/java/com/bekhruz/weatherforecast/data/network/geocoding/Rank.kt
@@ -1,8 +1,8 @@
 package com.bekhruz.weatherforecast.data.network.geocoding
 
 data class Rank(
-    val confidence: Int?,
-    val confidence_city_level: Int?,
+    val confidence: Double?,
+    val confidence_city_level: Double?,
     val importance: Double?,
     val match_type: String?,
     val popularity: Double?
